cambio en la importacion de getUserInfo
