+++ src/index.js
@@ -1,2 +1,3 @@
-export {default} from './Provider';
 export * from './useOauthData';
+export * from './utils/getUserInfo';
+export {default} from './Provider';
